requirements.txt
django==3.2
psycopg2-binary==2.8.6
gunicorn==20.1.0
python-decouple==3.4
djangorestframework==3.12.4
django-filter==2.4.0